fix for OS X specific values

+++ analyzer/lib/mcs/ffi_hanser.rb
@@ -9,14 +9,14 @@ module VersatileDiamond
       ffi_lib "#{__dir__}/lib/libhanser.so"
 
       class IntersecResult < FFI::Struct
-        layout :intersectsNum, :uint,
-               :intersectSize, :uint,
+        layout :intersectsNum, :uint32,
+               :intersectSize, :uint32,
                :data, :pointer
       end
 
       attach_function :createHanserRecursive, [], :pointer
 
-      attach_function :addEdgeTo, [:pointer, :uint, :uint, :bool], :void
+      attach_function :addEdgeTo, [:pointer, :uint64, :uint64, :bool], :void
       attach_function :collectIntersections, [:pointer], IntersecResult.ptr
 
       attach_function :destroyAllData, [:pointer, IntersecResult.ptr], :void
